新watcher权限管理

+++ watcher-manager/src/main/java/com/aihuishou/bi/vo/OperationUserVO.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Setter
 public class OperationUserVO {
     private String operation;
-    private List<Long> ids;
+    private List<Integer> ids;
 }
